Default datafile manager abstracted from HTTP request details
